requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-openai
langchain-classic
pandas